+++ admin/application/controllers/Policy.php
@@ -8,6 +8,11 @@ class policy extends loadFile
   {
     session_start();
     $this->db = $this->model('Model');
+
+    $all_permission_array = $this->permission();
+		$permition[] = $all_permission_array[2]->options;
+		$permition_option_array =explode(',',$permition[0]);
+		$this->_auth = $permition_option_array;
   }
 
   // To show all policies
@@ -18,14 +23,22 @@ class policy extends loadFile
     $option = '';
     $where = '';
     $records = $this->db->select_data($select, $tbl, $option, $where);
-    $this->view("managePolicy", array("title" => "Manage Policy", "policy_data" => $records));
+    if(in_array('1', $this->_auth)){  
+    $this->view("managePolicy", array("title" => "Manage Policy", "policy_data" => $records, 'recode'=> $this->_auth));
+    }else {
+     $this->single_view("error");
+}
   }
 
 
   // To show add policy page
   public function addPolicy()
   {
+    if(in_array('3', $this->_auth)){  
     $this->view("addPolicy", array("title" => "Manage Policy"));
+    }else {
+     $this->single_view("error");
+}
   }
 
   // To add new policy
@@ -92,7 +105,7 @@ class policy extends loadFile
   // To delete policy
   public function delPolicy()
   {
-    if (isset($_REQUEST['did'])) {
+    if ((isset($_REQUEST['did'])) && (in_array('4', $this->_auth))) {
       $id = $_REQUEST['did'];
       $wh = array("id" => $id);
       $this->db->delete_data("companypolicy", $wh);
@@ -109,7 +122,11 @@ class policy extends loadFile
     $option = '';
     $where = "id = $id";
     $records = $this->db->select_data($select, $tbl, $option, $where);
+    if(in_array('2', $this->_auth)){  
     $this->view("editPolicy", array("title" => "Update Your Policy", "policy_data" => $records));
+    }else {
+     $this->single_view("error");
+}
   }
 
   // To edit old policy
